third_party/blink/web_tests/fast/events/touch/gesture/gesture-tap-on-hover-element-expected.txt

Hover over me
Click me
Don't click me

Tests that a gesture tap is dispatched before the hover state on the active element is cleared.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

See https://bugs.webkit.org/show_bug.cgi?id=103283 for details.

The Click Me div should not be visible.
PASS clickMe.offsetTop is 0
The Click Me div should be visible after a GestureShowPress on the Hover Over Me div.
PASS hoverme is hoverme
PASS clickMe.offsetTop is 100
Tapping on the Click Me div should dispatch a click to the Click Me div and not to the Don't Click Me div and the Click Me div should still be visible.
PASS clickme is clickme
PASS clickMeCount is 1
PASS dontClickMeCount is 0
PASS clickMe.offsetTop is 100
PASS successfullyParsed is true

TEST COMPLETE
